+++ src/Resources/contao/languages/de/tl_content.php
@@ -8,14 +8,14 @@
  *
  * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
  *
- * last-updated: 2018-07-27T14:50:17+00:00
+ * last-updated: 2018-07-27T14:50:23+00:00
  */
-
 
 $GLOBALS['TL_LANG']['tl_content']['bs_grid']['0']          = 'Grid';
 $GLOBALS['TL_LANG']['tl_content']['bs_grid']['1']          = 'Bitte wählen Sie eine Grid-Definition.';
 $GLOBALS['TL_LANG']['tl_content']['bs_grid_name']['0']     = 'Grid-Name';
 $GLOBALS['TL_LANG']['tl_content']['bs_grid_name']['1']     = 'Interner Grid-Name für die Anzeige im Backend.';
+$GLOBALS['TL_LANG']['tl_content']['bs_grid_parent']['1']   = 'Bitte wählen Sie das Elternelement aus.';
 $GLOBALS['TL_LANG']['tl_content']['bs_grid_wizard']['0']   = 'Erstelle Trenner';
 $GLOBALS['TL_LANG']['tl_content']['bs_grid_wizard']['1']   = 'Bitte wählen Sie, wie viele Trenner automatisch erzeugt werden sollen.';
 $GLOBALS['TL_LANG']['tl_content']['bs_grid_wizard_legend'] = 'Assistent';