+++ src/Resources/contao/languages/de/tl_theme.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Translations are managed using Transifex. To create a new translation
+ * or to help to maintain an existing one, please register at transifex.com.
+ *
+ * @link https://www.transifex.com/signup/
+ * @link https://www.transifex.com/projects/p/contao-bootstrap/language/de/
+ *
+ * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
+ *
+ * last-updated: 2017-09-28T10:33:15+00:00
+ */
+
+
+$GLOBALS['TL_LANG']['tl_theme']['bootstrap_legend']     = 'Bootstrap';
+$GLOBALS['TL_LANG']['tl_theme']['bs_grid']['0']         = 'Spaltensets';
+$GLOBALS['TL_LANG']['tl_theme']['bs_grid']['1']         = 'Verwalten Sie Spaltensets.';
+$GLOBALS['TL_LANG']['tl_theme']['bs_grid_columns']['0'] = 'Grid-Spalten';
+$GLOBALS['TL_LANG']['tl_theme']['bs_grid_columns']['1'] = 'Bestimmen Sie die Anzahl der Spalten für Ihr Grid (Standard ist 12).';
+